Create basket view page, change router for product listing page
