+++ src/components/Pages/Basket.vue
@@ -1,0 +1,15 @@
+<template>
+  <div class="basket">
+    <h1>Your Basket</h1>
+    <!-- Add your basket logic and display here -->
+    <p>No items in your basket yet.</p>
+  </div>
+</template>
+
+<script setup lang="ts">
+// You can add your basket management logic here
+</script>
+
+<style scoped>
+/* Add any custom styles for your basket view */
+</style>
