+++ src/router.ts
@@ -1,13 +1,19 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import App from './App.vue'
+import Products from './components/Pages/Products.vue'
+import Basket from './components/Pages/Basket.vue'
 
 const routes = [
   {
     path: '/',
     name: 'Product listing',
-    component: App,
+    component: Products,
     // We'll use a query parameter `page` to handle the pagination.
     props: (route) => ({ page: parseInt(route.query.page as string) || 1 }),
+  },
+  {
+    path: '/basket',
+    name: 'Basket',
+    component: Basket, // Set BasketView as the component for this route
   },
 ]
